pyproject.toml
[tool.poetry]
name = "tag-by-target"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "tag_by_target"}]

[tool.poetry.dependencies]
python = "3.9.10"
pysnyk = "^0.9.0"


[tool.poetry.group.dev.dependencies]
mypy = "^0.931"
black = "^22.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
